backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for AI Block Software Development"
authors = [{name = "Worawat Lawanont", email = "worawat.law@gmail.com"}]
dependencies = [
    "Django >=3.2",
    "djangorestframework >=3.12",
    "django-cors-headers >=3.7",
    "requests >=2.25"
]

[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for AI Block Software Development"
authors = ["worawat.law@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12"
django-cors-headers = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-django = "^4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
